dodano walidację do deckController, stworzono cardController

+++ src/main/java/controllers/DeckController.java
@@ -1,10 +1,18 @@
 package controllers;
 
+import database.document.CardDocument;
 import database.document.DeckDocument;
+import database.repository.CardRepository;
 import database.repository.DeckRepository;
+
+import java.util.List;
+
 import static java.net.HttpURLConnection.HTTP_BAD_REQUEST;
+import static java.net.HttpURLConnection.HTTP_INTERNAL_ERROR;
 import static java.net.HttpURLConnection.HTTP_OK;
 import static spark.Spark.*;
+import static database.mongo.MongoUtils.toJson;
+
 
 /**
  * Created by Andrzej on 2016-11-09.
@@ -17,7 +25,7 @@ public class DeckController extends AbstractController {
         this.repository = new DeckRepository();
     }
 
-    public void setRestApi(){
+    public void registerRestApi(){
         post("/decks/create", (req, resp)-> {
             String owner = req.queryParams("ownerId");
             String name = req.queryParams("name");
@@ -33,21 +41,43 @@ public class DeckController extends AbstractController {
                 }
             } catch (NumberFormatException exception){
                 resp.status(HTTP_BAD_REQUEST);
+            }catch (Exception  e){
+                resp.status(HTTP_INTERNAL_ERROR);
             }
 
             return response;
         });
 
         delete("/decks/delete/:id", (req, resp) -> {
+            String result = "";
             String id = req.params("id");
-            Object deck = repository.delete(id);
-            if(deck != null){
-                resp.status(HTTP_OK);
-            }
-            else{
+            if(id!=null && !id.isEmpty()){
+                try{
+                    Object deck = repository.delete(id);
+                    CardRepository cardRepository = new CardRepository();
+                    List<CardDocument> cards = cardRepository.findByDeckId(id);
+                    for (CardDocument card: cards) {
+                        cardRepository.delete(card);
+                    }
+
+                    if(deck != null){
+                        resp.status(HTTP_OK);
+                    }
+                    else{
+                        resp.status(HTTP_BAD_REQUEST);
+                    }
+                }catch (Exception e){
+                    resp.status(HTTP_INTERNAL_ERROR);
+                    result = e.getMessage();
+                }
+            }else {
                 resp.status(HTTP_BAD_REQUEST);
+                result = "Deck Id can not be null";
             }
-            return "";
+
+            return result;
         });
+
+        get("/decks", (request, response) -> toJson(repository.findAll()));
     }
 }
